Three tier approach applied to inductively loaded dickson
The three tier approach is applied to the inductily loaded dickson. In its current form the state sensitivity approach oscillates between two solutions. The code is unable to separate diode conduction ON times after the first iteration. A solution to this issue was implemented by inserting an exit condition to use time stepping approach to correct the diode timing.